+++ lib/constants/index.ts
@@ -1,0 +1,13 @@
+/**
+ * constantes do site mimo - re-exporta todos os módulos.
+ * 
+ * este arquivo centraliza todas as exportações para manter compatibilidade
+ * com imports existentes que usam '@/lib/constants'.
+ */
+
+export * from './contact'
+export * from './home'
+export * from './services'
+export * from './celebrities'
+export * from './jobs'
+